seperate time since posted functionality into its own component
